+++ src/main/java/br/com/zalf/prolog/webservice/BuildConfig.java
@@ -2,5 +2,5 @@ package br.com.zalf.prolog.webservice;
 
 public class BuildConfig {
 	 public static final boolean DEBUG = true;
-	 public static final String VERSION_CODE = "v2_1_0_2";
+	 public static final String VERSION_CODE = "v2_1_0_3";
 }